refactor: Move AppLike to test folder
